Move OpenVPN/WireGuard l10n to respective extensions
