+++ Passepartout/AppShared/L10n/TunnelKit+L10n.swift
@@ -107,104 +107,3 @@ extension IPv6Settings.Route {
         "\(destination)/\(prefixLength) -> \(gateway ?? "*")"
     }
 }
-
-extension Error {
-    var localizedVPNDescription: String? {
-        if let ovpnError = self as? OpenVPNProviderError {
-            return ovpnErrorDescription(ovpnError)
-        }
-        if let wgError = self as? WireGuardProviderError {
-            return wgErrorDescription(wgError)
-        }
-        if let neError = self as? NEVPNError {
-            return neErrorDescription(neError)
-        }
-        return localizedDescription
-    }
-
-    private func ovpnErrorDescription(_ error: OpenVPNProviderError) -> String? {
-        let V = L10n.Tunnelkit.Errors.Vpn.self
-        switch error {
-        case .socketActivity, .timeout:
-            return V.timeout
-
-        case .dnsFailure:
-            return V.dns
-
-        case .tlsInitialization, .tlsServerVerification, .tlsHandshake:
-            return V.tls
-
-        case .authentication:
-            return V.auth
-
-        case .encryptionInitialization, .encryptionData:
-            return V.encryption
-
-        case .serverCompression, .lzo:
-            return V.compression
-
-        case .networkChanged:
-            return V.network
-
-        case .routing:
-            return V.routing
-
-        case .gatewayUnattainable:
-            return V.gateway
-
-        case .serverShutdown:
-            return V.shutdown
-
-        default:
-            return nil
-        }
-    }
-
-    private func wgErrorDescription(_ error: WireGuardProviderError) -> String? {
-        let V = L10n.Tunnelkit.Errors.Vpn.self
-        switch error {
-        case .dnsResolutionFailure:
-            return V.dns
-
-        default:
-            return nil
-        }
-    }
-
-    private func neErrorDescription(_ error: NEVPNError) -> String? {
-        error.localizedDescription.capitalized
-    }
-}
-
-extension OpenVPN.ConfigurationError: LocalizedError {
-    public var errorDescription: String? {
-        let V = L10n.Tunnelkit.Errors.Openvpn.self
-        switch self {
-        case .encryptionPassphrase:
-            pp_log.error("Could not parse configuration URL: unable to decrypt, passphrase required")
-            return V.passphraseRequired
-
-        case .unableToDecrypt(let error):
-            pp_log.error("Could not parse configuration URL: unable to decrypt, \(error.localizedDescription)")
-            return V.decryption
-
-        case .malformed(let option):
-            pp_log.error("Could not parse configuration URL: malformed option, \(option)")
-            return V.malformed(option)
-
-        case .missingConfiguration(let option):
-            pp_log.error("Could not parse configuration URL: missing configuration, \(option)")
-            return V.requiredOption(option)
-
-        case .unsupportedConfiguration(var option):
-            if option.contains("external") {
-                option.append(" (see FAQ)")
-            }
-            pp_log.error("Could not parse configuration URL: unsupported configuration, \(option)")
-            return V.unsupportedOption(option)
-        }
-    }
-}
-
-extension WireGuard.ConfigurationError: LocalizedError {
-}
