+++ Passepartout/AppShared/L10n/WireGuard+L10n.swift
@@ -24,4 +24,20 @@
 //
 
 import Foundation
-import TunnelKitWireGuardCore
+import TunnelKitWireGuard
+
+extension WireGuardProviderError: LocalizedError {
+    public var errorDescription: String? {
+        let V = L10n.Tunnelkit.Errors.Vpn.self
+        switch self {
+        case .dnsResolutionFailure:
+            return V.dns
+
+        default:
+            return nil
+        }
+    }
+}
+
+extension WireGuard.ConfigurationError: LocalizedError {
+}
